+++ src/consts.rs
@@ -77,10 +77,6 @@ pub const METADATA_SYMBOL: &str = "ORE";
 /// The uri for token metdata.
 pub const METADATA_URI: &str = "https://ore.supply/metadata.json";
 
-/// Program id of the compute budge program.
-pub const COMPUTE_BUDGET_PROGRAM_ID: Pubkey =
-    pubkey!("ComputeBudget111111111111111111111111111111");
-
 /// Program id for const pda derivations
 const PROGRAM_ID: [u8; 32] = unsafe { *(&crate::id() as *const Pubkey as *const [u8; 32]) };
