+++ src/lib.rs
@@ -1,3 +1,4 @@
+pub mod builders;
 pub mod consts;
 pub mod error;
 pub mod instruction;
@@ -14,7 +15,6 @@ use solana_program::{
     program_error::ProgramError, pubkey::Pubkey,
 };
 
-// TODO Close proof accounts to recover sol
 // TODO Is downgrade necessary?
 
 declare_id!("mineRHF5r6S7HyD9SppBfVMXMavDkJsxwGesEvxZr2A");
@@ -36,13 +36,13 @@ pub fn process_instruction(
         .ok_or(ProgramError::InvalidInstructionData)?;
 
     match OreInstruction::try_from(*tag).or(Err(ProgramError::InvalidInstructionData))? {
-        OreInstruction::Reset => process_reset(program_id, accounts, data)?,
         OreInstruction::Register => process_register(program_id, accounts, data)?,
+        OreInstruction::Deregister => process_deregister(program_id, accounts, data)?,
+        OreInstruction::Reset => process_reset(program_id, accounts, data)?,
         OreInstruction::Mine => process_mine(program_id, accounts, data)?,
         OreInstruction::Claim => process_claim(program_id, accounts, data)?,
         OreInstruction::Stake => process_stake(program_id, accounts, data)?,
         OreInstruction::Upgrade => process_upgrade(program_id, accounts, data)?,
-        OreInstruction::Deregister => process_deregister(program_id, accounts, data)?,
         OreInstruction::Initialize => process_initialize(program_id, accounts, data)?,
         OreInstruction::UpdateAdmin => process_update_admin(program_id, accounts, data)?,
         OreInstruction::UpdateTolerance => process_update_tolerance(program_id, accounts, data)?,
