+++ src/processor/deregister.rs
@@ -1,23 +1,14 @@
 use solana_program::{
-    account_info::AccountInfo, entrypoint::ProgramResult, program_error::ProgramError,
-    pubkey::Pubkey, system_program,
+    account_info::AccountInfo, entrypoint::ProgramResult, program::invoke_signed,
+    program_error::ProgramError, pubkey::Pubkey, system_program,
 };
 
-use crate::{instruction::RegisterArgs, loaders::*};
+use crate::{loaders::*, state::Proof, utils::AccountDeserialize, PROOF};
 
-/// Register generates a new hash chain for a prospective miner. Its responsibilities include:
-/// 1. Initialize a new proof account.
-/// 2. Generate an initial hash from the signer's key.
-///
-/// Safety requirements:
-/// - Register is a permissionless instruction and can be invoked by any singer.
-/// - Can only succeed if the provided proof acount PDA is valid (associated with the signer).
-/// - Can only succeed once per signer.
-/// - The provided system program must be valid.
 pub fn process_deregister<'a, 'info>(
     _program_id: &Pubkey,
     accounts: &'a [AccountInfo<'info>],
-    data: &[u8],
+    _data: &[u8],
 ) -> ProgramResult {
     // Load accounts
     let [signer, proof_info, system_program] = accounts else {
@@ -27,10 +18,37 @@ pub fn process_deregister<'a, 'info>(
     load_proof(proof_info, signer.key, true)?;
     load_program(system_program, system_program::id())?;
 
-    // TODO Ensure proof.balance == 0
-    // TODO Send lamports to signer
-    // TODO Realloc data to 0
-    // TODO Reassign back to system program
+    // Validate balance is zero
+    let proof_data = proof_info.data.borrow();
+    let proof = Proof::try_from_bytes(&proof_data)?;
+    if proof.balance.gt(&0) {
+        return Err(ProgramError::InvalidAccountData);
+    }
+
+    // Generate bump
+    let bump = Pubkey::find_program_address(&[PROOF, signer.key.as_ref()], &crate::id()).1;
+
+    // Realloc data to zero
+    drop(proof_data);
+    proof_info.realloc(0, false)?;
+
+    // Send lamports to signer
+    invoke_signed(
+        &solana_program::system_instruction::transfer(
+            proof_info.key,
+            signer.key,
+            proof_info.lamports(),
+        ),
+        &[proof_info.clone(), signer.clone(), system_program.clone()],
+        &[&[PROOF, signer.key.as_ref(), &[bump]]],
+    )?;
+
+    // Reassign back to system program
+    solana_program::program::invoke_signed(
+        &solana_program::system_instruction::assign(proof_info.key, &system_program::id()),
+        &[proof_info.clone(), system_program.clone()],
+        &[&[PROOF, signer.key.as_ref(), &[bump]]],
+    )?;
 
     Ok(())
 }
